Fix caching of requests with Set-Cookie
Odoo publishes static files with a Set-Cookie header that contains the
session of the user. The cache entry was ignoring the Set-Cookie header
for the key of the entry, but was returning it in the cached response.
Which means that the first user to login would fill the cache with their
cookie. The subsequent users to reach the nginx cache (users having a
browser cache should not be affected) would receive a request containing
a Set-Cookie header of the first user → the browser obtains the
session_id of another user.

The reason was:
* proxy_ignore_headers removes the Set-Cookie from the cache entry key,
but not from the cached response
* proxy_hide_header hides the header in the cached response

Alongside this change, 'proxy_set_header Cookie ""' prevents to pass any
cookie to odoo on static files.