Remove forwarding of If-None-Match
When we forward the header, Odoo will (legitimately) return 304
responses without content, which will get cached by nginx as such.
Opening odoo from a fresh browser (with no prior files cached on disk)
will receive a 304 with no content.

There may be a way exclude this behavior from the cache, but the most
effective and safe change is to remove the header when forwarding to
odoo. Odoo will always have to read and return the whole content and
will have no possibility to return a 304 based on the If-None-Match
Etag, but this is the responsibility of nginx.

I tried this tuning that did not help:

      proxy_cache_valid 200 301 302 60m;
      proxy_cache_valid 304 0;
      proxy_cache_valid 404 0;