[FIX] raise proxy_buffers and proxy_buffer_size
On Odoo 17 there has been a change in the hearders for the PDF viewer,
which cause a larger header to be sent by Odoo on the response the the HTTP
request for a document preview.

This larger header causes an error in Nginx because the value of
is too small to hold it.

We raise that value from 1k to 4k.

We also raise the value of proxy_buffers: the memory used by nginx is well
inside the limits of the pod in k8s so we can get a little bit more ease
in there.

closes: BGIS-1374